Avoid code duplication with LinkSerializationHelper
